#912 Change: Extract notification center
