+++ src/components/product/detail/dealer/__tests__/ProductDealerItems.test.tsx
@@ -25,7 +25,7 @@ vi.mock("embla-carousel-react", () => ({
     ],
 }));
 
-vi.mock("@/hooks/notification/useMarkNotificationSeen.ts", () => ({
+vi.mock("@/features/notification-center/api/useMarkNotificationSeen.ts", () => ({
     useMarkNotificationSeen: () => ({ mutate: vi.fn() }),
 }));
