+++ src/components/product/grid/__tests__/ProductGridItem.test.tsx
@@ -6,7 +6,7 @@ import { ProductGridItem } from "../ProductGridItem.tsx";
 
 const mockMutate = vi.fn();
 
-vi.mock("@/hooks/notification/useMarkNotificationSeen.ts", () => ({
+vi.mock("@/features/notification-center/api/useMarkNotificationSeen.ts", () => ({
     useMarkNotificationSeen: () => ({ mutate: mockMutate }),
 }));
