+++ src/components/product/overview/__tests__/ProductCard.test.tsx
@@ -6,7 +6,7 @@ import { vi } from "vitest";
 
 const mockMutate = vi.fn();
 
-vi.mock("@/hooks/notification/useMarkNotificationSeen.ts", () => ({
+vi.mock("@/features/notification-center/api/useMarkNotificationSeen.ts", () => ({
     useMarkNotificationSeen: () => ({ mutate: mockMutate }),
 }));
