+++ src/features/notification-center/api/__tests__/useDeleteAllNotifications.test.ts
