+++ src/features/notification-center/api/__tests__/useMarkAllNotificationsSeen.test.ts
