+++ src/features/notification-center/api/__tests__/useMarkNotificationSeen.test.ts
