+++ src/features/notification-center/api/useDeleteAllNotifications.ts
