+++ src/features/notification-center/api/useMarkAllNotificationsSeen.ts
