+++ src/features/notification-center/api/useMarkNotificationSeen.ts
