+++ src/routes/$lng._auth.me.notifications.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { NotificationResults } from "@/components/notification/NotificationResults.tsx";
+import { NotificationCenterPage } from "@/features/notification-center/pages/NotificationCenterPage.tsx";
 import { generatePageHeadMeta } from "@/lib/seo/pageHeadMeta.ts";
 
 export const Route = createFileRoute("/$lng/_auth/me/notifications")({
@@ -10,7 +10,7 @@ export const Route = createFileRoute("/$lng/_auth/me/notifications")({
 function RouteComponent() {
     return (
         <div className="max-w-6xl mx-auto flex flex-col gap-8 pt-8 pb-8 ml-8 mr-8 lg:ml-auto lg:mr-auto">
-            <NotificationResults />
+            <NotificationCenterPage />
         </div>
     );
 }
